+++ nonebot_plugin_datastore/script/cli.py
@@ -1,14 +1,35 @@
 from argparse import Namespace
-from typing import Optional
+from functools import partial, wraps
+from typing import Any, Callable, Coroutine, Optional, TypeVar
 
 import anyio
 import click
 from nonebot.log import logger
+from typing_extensions import ParamSpec
 
 from ..config import plugin_config
 from ..plugin import PluginData
 from . import command
-from .utils import Config, get_plugins, run_async, run_sync
+from .utils import Config, get_plugins
+
+P = ParamSpec("P")
+R = TypeVar("R")
+
+
+def run_sync(func: Callable[P, R]) -> Callable[P, Coroutine[Any, Any, R]]:
+    @wraps(func)
+    async def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
+        return await anyio.to_thread.run_sync(partial(func, *args, **kwargs))
+
+    return wrapper
+
+
+def run_async(func: Callable[P, Coroutine[Any, Any, R]]) -> Callable[P, R]:
+    @wraps(func)
+    def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
+        return anyio.from_thread.run(partial(func, *args, **kwargs))
+
+    return wrapper
 
 
 @click.group()
