+++ nonebot_plugin_datastore/script/utils.py
@@ -1,14 +1,11 @@
-from functools import partial, wraps
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Coroutine, List, Optional, TypeVar
+from typing import TYPE_CHECKING, List, Optional
 
-import anyio
 from alembic import context
 from alembic.config import Config as AlembicConfig
 from click import BadParameter
 from nonebot import get_loaded_plugins, get_plugin
 from nonebot.log import logger
-from typing_extensions import ParamSpec
 
 from ..db import get_engine
 from ..plugin import PluginData
@@ -121,23 +118,3 @@ async def run_upgrade():
         config = Config(plugin)
         # 升级到最新版本
         await upgrade(config, "head")
-
-
-P = ParamSpec("P")
-R = TypeVar("R")
-
-
-def run_sync(func: Callable[P, R]) -> Callable[P, Coroutine[Any, Any, R]]:
-    @wraps(func)
-    async def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
-        return await anyio.to_thread.run_sync(partial(func, *args, **kwargs))
-
-    return wrapper
-
-
-def run_async(func: Callable[P, Coroutine[Any, Any, R]]) -> Callable[P, R]:
-    @wraps(func)
-    def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
-        return anyio.from_thread.run(partial(func, *args, **kwargs))
-
-    return wrapper
