+++ pyproject.toml
@@ -19,6 +19,7 @@ alembic = "^1.9.1"
 
 anyio = { version = ">=3.6", optional = true }
 click = { version = ">=8.0", optional = true }
+typing-extensions = { version = ">=4.4", optional = true }
 
 [tool.poetry.group.dev.dependencies]
 black = "*"
@@ -33,7 +34,7 @@ nb-cli = "^1.0.0"
 nonebot2 = { extras = ["httpx", "fastapi"], version = "^2.0.0-rc.1" }
 
 [tool.poetry.extras]
-cli = ["anyio", "click"]
+cli = ["anyio", "click", "typing-extensions"]
 
 [tool.black]
 line-length = 88
